+++ example/lib/menu_page.dart
@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:maplibre_example/annotations_page.dart';
 import 'package:maplibre_example/callbacks_page.dart';
 import 'package:maplibre_example/controller_page.dart';
+import 'package:maplibre_example/events_page.dart';
 import 'package:maplibre_example/flutter_map_page.dart';
 import 'package:maplibre_example/layers_circle_page.dart';
 import 'package:maplibre_example/layers_fill_extrusion_page.dart';
@@ -11,6 +12,7 @@ import 'package:maplibre_example/layers_heatmap_page.dart';
 import 'package:maplibre_example/layers_hillshade_page.dart';
 import 'package:maplibre_example/layers_line_page.dart';
 import 'package:maplibre_example/layers_raster_page.dart';
+import 'package:maplibre_example/layers_symbol_page.dart';
 import 'package:maplibre_example/styled_map_page.dart';
 import 'package:maplibre_example/two_maps_page.dart';
 import 'package:maplibre_example/web_controls_page.dart';
@@ -50,6 +52,11 @@ class MenuPage extends StatelessWidget {
                 label: 'Callbacks',
                 iconData: Icons.back_hand,
                 location: CallbacksPage.location,
+              ),
+              ItemCard(
+                label: 'Events',
+                iconData: Icons.notifications,
+                location: EventsPage.location,
               ),
               if (kIsWeb)
                 ItemCard(
@@ -108,6 +115,11 @@ class MenuPage extends StatelessWidget {
                 label: 'Raster',
                 iconData: Icons.grid_on,
                 location: LayersRasterPage.location,
+              ),
+              ItemCard(
+                label: 'Symbol',
+                iconData: Icons.emoji_emotions,
+                location: LayersSymbolPage.location,
               ),
             ],
           ),